requirements-opt.txt
django-simple-captcha
djangorestframework
django-cors-headers
django-compressor
django-libsass
python-memcached
psycopg2